Allow specifying an interface to bind to:
Users with multiple interfaces can limit listening
to a single interface.
